+++ src/DevilDaggersInfo.Tools/Ui/Main/MainWindow.cs
@@ -1,6 +1,7 @@
 using DevilDaggersInfo.Tools.Engine.Maths.Numerics;
 using DevilDaggersInfo.Tools.Ui.CustomLeaderboards.LeaderboardList;
 using DevilDaggersInfo.Tools.Ui.ModManager;
+using DevilDaggersInfo.Tools.Ui.ModManager.ModsDirectory;
 using DevilDaggersInfo.Tools.Utils;
 using ImGuiNET;
 using System.Numerics;
@@ -74,7 +75,7 @@ public static class MainWindow
 				static void GoToModManager()
 				{
 					UiRenderer.Layout = LayoutType.ModManager;
-					ModsDirectoryWindow.LoadModsDirectory();
+					ModsDirectoryWindow.Initialize();
 				}
 
 				static void GoToCustomLeaderboards()
