+++ src/DevilDaggersInfo.Tools/Ui/ModManager/ModManagerWindow.cs
@@ -1,3 +1,4 @@
+using DevilDaggersInfo.Tools.Ui.ModManager.ModsDirectory;
 using ImGuiNET;
 using System.Numerics;
