+++ src/DevilDaggersInfo.Tools/Ui/ModManager/ModsDirectory/Data/ModFile.cs
@@ -1,0 +1,5 @@
+using DevilDaggersInfo.Core.Mod;
+
+namespace DevilDaggersInfo.Tools.Ui.ModManager.ModsDirectory.Data;
+
+public sealed record ModFile(string FileName, ModFileType FileType, ModBinaryType? Type, int? ChunkCount, long FileSize);
