+++ src/DevilDaggersInfo.Tools/Ui/ModManager/ModsDirectory/Data/ModFileType.cs
@@ -1,0 +1,10 @@
+namespace DevilDaggersInfo.Tools.Ui.ModManager.ModsDirectory.Data;
+
+public enum ModFileType
+{
+	EnabledMod,
+	DisabledMod,
+	ModWithInvalidPrefix,
+	Other,
+	Error,
+}
