+++ src/DevilDaggersInfo.Tools/Ui/ModManager/ModsDirectory/ModsDirectoryWindow.cs
@@ -1,29 +1,24 @@
-using DevilDaggersInfo.Core.Mod;
-using DevilDaggersInfo.Core.Mod.Exceptions;
-using DevilDaggersInfo.Tools.Ui.Popups;
+using DevilDaggersInfo.Tools.Ui.ModManager.ModsDirectory.Data;
 using DevilDaggersInfo.Tools.User.Settings;
 using DevilDaggersInfo.Tools.Utils;
 using ImGuiNET;
 using System.Numerics;
 
-namespace DevilDaggersInfo.Tools.Ui.ModManager;
+namespace DevilDaggersInfo.Tools.Ui.ModManager.ModsDirectory;
 
 public static class ModsDirectoryWindow
 {
-	private static List<ModFile> _modFiles = new();
+	private static readonly ModsDirectoryLogic _logic = new();
+
 	private static bool _showEnabledMods = true;
 	private static bool _showDisabledMods = true;
 	private static bool _showModsWithInvalidPrefix = true;
 	private static bool _showOtherFiles;
 	private static bool _showErrors = true;
 
-	private enum ModFileType
+	public static void Initialize()
 	{
-		EnabledMod,
-		DisabledMod,
-		ModWithInvalidPrefix,
-		Other,
-		Error,
+		_logic.LoadModsDirectory();
 	}
 
 	public static void Render()
@@ -39,7 +34,7 @@ public static class ModsDirectoryWindow
 			"""));
 
 		if (ImGui.Button("Reload"))
-			LoadModsDirectory();
+			_logic.LoadModsDirectory();
 
 		ImGui.Separator();
 
@@ -90,7 +85,7 @@ public static class ModsDirectoryWindow
 					0 => "The name of the file in the mods folder",
 					1 => "The mod type (audio or dd, core mods are not supported by the game)",
 					2 => "The chunk count is typically the amount of game assets changed by the mod",
-					3 => "The file size in bytes",
+					3 => "The mod's file size",
 					_ => string.Empty,
 				};
 				ImGui.SetTooltip(tooltip);
@@ -102,21 +97,14 @@ public static class ModsDirectoryWindow
 				uint sorting = sortsSpecs.Specs.ColumnUserID;
 				bool sortAscending = sortsSpecs.Specs.SortDirection == ImGuiSortDirection.Ascending;
 
-				_modFiles = sorting switch
-				{
-					0 => sortAscending ? _modFiles.OrderBy(m => m.FileName).ToList() : _modFiles.OrderByDescending(m => m.FileName).ToList(),
-					1 => sortAscending ? _modFiles.OrderBy(m => m.Type).ToList() : _modFiles.OrderByDescending(m => m.Type).ToList(),
-					2 => sortAscending ? _modFiles.OrderBy(m => m.ChunkCount).ToList() : _modFiles.OrderByDescending(m => m.ChunkCount).ToList(),
-					3 => sortAscending ? _modFiles.OrderBy(m => m.FileSize).ToList() : _modFiles.OrderByDescending(m => m.FileSize).ToList(),
-					_ => throw new InvalidOperationException($"Invalid sorting column '{sorting}'."),
-				};
+				_logic.SortModFiles(sorting, sortAscending);
 
 				sortsSpecs.SpecsDirty = false;
 			}
 
-			for (int i = 0; i < _modFiles.Count; i++)
+			for (int i = 0; i < _logic.ModFiles.Count; i++)
 			{
-				ModFile modFile = _modFiles[i];
+				ModFile modFile = _logic.ModFiles[i];
 
 				if (!_showEnabledMods && modFile.FileType == ModFileType.EnabledMod)
 					continue;
@@ -161,58 +149,6 @@ public static class ModsDirectoryWindow
 		ImGui.Text(span);
 	}
 
-	public static void LoadModsDirectory()
-	{
-		try
-		{
-			_modFiles.Clear();
-
-			string[] files = Directory.GetFiles(UserSettings.ModsDirectory);
-			foreach (string file in files)
-			{
-				_modFiles.Add(ReadModFile(file));
-			}
-
-			_modFiles = _modFiles.OrderBy(m => m.FileName).ToList();
-		}
-		catch (Exception ex)
-		{
-			PopupManager.ShowError("Error loading files in the mods directory.\n\n" + ex.Message);
-			Root.Log.Error(ex, "Error loading files in the mods directory.");
-		}
-	}
-
-	private static ModFile ReadModFile(string filePath)
-	{
-		string fileName = Path.GetFileName(filePath);
-
-		long fileSize = 0;
-		try
-		{
-			using FileStream fs = new(filePath, FileMode.Open);
-			fileSize = fs.Length;
-			using BinaryReader reader = new(fs);
-			ModBinaryToc modBinaryToc = ModBinaryToc.FromReader(reader);
-
-			ModFileType fileType =
-				fileName.StartsWith("audio") || fileName.StartsWith("dd") ? ModFileType.EnabledMod :
-				fileName.StartsWith("_audio") || fileName.StartsWith("_dd") ? ModFileType.DisabledMod :
-				ModFileType.ModWithInvalidPrefix;
-
-			return new(fileName, fileType, modBinaryToc.Type, modBinaryToc.Chunks.Count, fileSize);
-		}
-		catch (InvalidModBinaryException)
-		{
-			return new(fileName, ModFileType.Other, null, null, fileSize);
-		}
-		catch (Exception ex)
-		{
-			PopupManager.ShowError($"Error loading file '{filePath}'.\n\n" + ex.Message);
-			Root.Log.Error(ex, $"Error loading file '{filePath}'.");
-			return new(fileName, ModFileType.Error, null, null, fileSize);
-		}
-	}
-
 	private static Vector4 GetColor(ModFileType modFileType)
 	{
 		return modFileType switch
@@ -225,6 +161,4 @@ public static class ModsDirectoryWindow
 			_ => new(1),
 		};
 	}
-
-	private sealed record ModFile(string FileName, ModFileType FileType, ModBinaryType? Type, int? ChunkCount, long FileSize);
 }
